minor updates for linux

+++ ex00/ConversionsFunctions.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ConversionsFunctions.cpp                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vadimhrabrov <vadimhrabrov@student.42.f    +#+  +:+       +#+        */
+/*   By: vkhrabro <vkhrabro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/22 21:22:33 by vkhrabro          #+#    #+#             */
-/*   Updated: 2024/09/24 22:31:20 by vadimhrabro      ###   ########.fr       */
+/*   Updated: 2024/09/26 19:12:52 by vkhrabro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ void ScalarConverter::convertToInt(T input) {
     // Check for NaN or infinity
     if (std::isnan(input) || std::isinf(input))
         std::cout << "int: impossible" << std::endl;
-    else if (input > INT_MAX || input < INT_MIN)
+    else if (input > static_cast<float>(INT_MAX) || input < static_cast<float>(INT_MIN))
         std::cout << "int: impossible" << std::endl;
     else
         std::cout << "int: " << static_cast<int>(input) << std::endl;
